smart_study_assistant/requirements.txt
langchain
chromadb
huggingface_hub
openai
python-dotenv

